Fixed a security flaw in OAuth handling.
FF's SP returns 'username' parameter as a response to Access Token
request. HTTP response in OAuth Core 1.0{,a} does not signed so we must
not rely on that value. To use OAuth as an authentication scheme in
addition to authorization scheme, we must ask to SP who owns the access
token. Fixed.